Struct sscs::World

pub struct World<T, C> {
    pub component_data: T,
    pub global_data: C,
    // some fields omitted
}

This struct holds everything related to entity-component system.

Fields

component_data

Holds all the components attached to entities.

global_data

Holds all data that isn't related to any entit.y

Methods

impl<T: Components, C: Encodable + Decodable> World<T, C>

fn new(global_data: C) -> World<T, C>

Creates a new World

fn add_entity(&mut self) -> Entity

Adds a new entity with no components

fn delete_entity(&mut self, e: &Entity)

Marks an entity for deletion. The entity gets actually deleted the next time you call update

fn entity_valid(&self, e: &Entity) -> bool

Check if an entity actually exists.

fn entities_with_components(&self, mask: u32) -> Vec<Entity>

Returns all entities that have at least the components specified by mask

fn entities(&self) -> Vec<Entity>

Returns all valid entities

fn get_component_mask(&self, entity: &Entity) -> u32

fn has<Z>(&self, entity: &Entity) -> bool where Z: ComponentAccess<T>

Checks if entity has a component of type Z

fn get<Z>(&self, entity: &Entity) -> Option<&Z> where Z: ComponentAccess<T>

Returns a reference to component Z attached to `entity

fn get_mut<Z>(&mut self, entity: &Entity) -> Option<&mut Z> where Z: ComponentAccess<T>

Returns a mutable reference to component Z attached to entity

fn add<Z>(&mut self, entity: &Entity, comp: Z) where Z: ComponentAccess<T>

Adds a component of type Z to entity If entity already has component of this type, it gets overwritten.

fn remove<Z>(&mut self, entity: &Entity) where Z: ComponentAccess<T>

Removes a component of type Z from entity If entity doesn't own component Z, this function does nothing.

fn update(&mut self, systems: &mut Vec<&mut System<Self>>)

Removes entities marked for deletion and runs systems.