Trait sscs::ComponentAccess

pub trait ComponentAccess<T>: Sized {
    fn mask() -> u32;
    fn get_data(comps: &T) -> &Vec<Self>;
    fn get_data_mut(comps: &mut T) -> &mut Vec<Self>;
}

Trait for components access

Required Methods

fn mask() -> u32

Returns the mask used to check component ownership

fn get_data(comps: &T) -> &Vec<Self>

Returns a reference to the vector holding all components of this type

fn get_data_mut(comps: &mut T) -> &mut Vec<Self>

Returns a mutable reference to the vector holding all components of this type