Trait sscs::System

pub trait System<W> {
    fn process(&mut self, entities: Vec<Entity>, world: &mut W);
    fn get_entity_mask(&self) -> u32;

    fn process_added(&mut self, _: Vec<Entity>, _: &mut W) { ... }
    fn process_removed(&mut self, _: Vec<Entity>, _: &mut W) { ... }
}

Trait for systems

Required Methods

fn process(&mut self, entities: Vec<Entity>, world: &mut W)

Function that processes all the entities that have the components specified by get_entity_mask.

fn get_entity_mask(&self) -> u32

Returns a bitmask that represents the minimum combination of components an entity needs to have in order to be processed by this system Needs to be constant

Provided Methods

fn process_added(&mut self, _: Vec<Entity>, _: &mut W)

Function for processing all the entities that got added to this system last frame. Entities added/removed here will get deleted next frame.

fn process_removed(&mut self, _: Vec<Entity>, _: &mut W)

Function for processing all the entities that got removed from this system last frame. Entities added/removed here will get deleted next frame.